[Enhancement](paimon)support native read paimon top level schema change table. (#48723)
Problem Summary:
Supports native reader reading tables after the top-level schema of
paimon is changed, but does not support tables after the internal schema
of struct is changed.

change  top-level schema(support):
```sql
--spark sql
ALTER TABLE table_name ADD COLUMNS (c1 INT,c2 STRING);
ALTER TABLE table_name RENAME COLUMN c0 TO c1;
ALTER TABLE table_name DROP COLUMNS (c1, c2);
ALTER TABLE table_name ADD COLUMN c INT FIRST;
ALTER TABLE table_name ADD COLUMN c INT AFTER b;
ALTER TABLE table_name ALTER COLUMN col_a FIRST;
ALTER TABLE table_name ALTER COLUMN col_a AFTER col_b;
```

change internal schema of struct schema(not support, will support in the
next PR):
```sql
--spark sql
ALTER TABLE table_name ADD COLUMN        v.value.f3 STRING;
ALTER TABLE table_name RENAME COLUMN v.f1 to f100;
ALTER TABLE table_name DROP COLUMN      v.value.f3 ;
ALTER TABLE table_name ALTER COLUMN      v.col_a FIRST;
```

+++ be/src/vec/exec/format/table/iceberg_reader.h
@@ -148,7 +148,7 @@ protected:
     // copy from _colname_to_value_range with new column name that is in parquet/orc file, to support schema evolution.
     std::unordered_map<std::string, ColumnValueRangeType> _new_colname_to_value_range;
     // column id to name map. Collect from FE slot descriptor.
-    std::unordered_map<int, std::string> _col_id_name_map;
+    std::unordered_map<uint64_t, std::string> _col_id_name_map;
     // col names in the parquet,orc file
     std::vector<std::string> _all_required_col_names;
     // col names in table but not in parquet,orc file
@@ -194,7 +194,7 @@ public:
                                  kv_cache, io_ctx) {}
     Status init_reader(
             const std::vector<std::string>& file_col_names,
-            const std::unordered_map<int, std::string>& col_id_name_map,
+            const std::unordered_map<uint64_t, std::string>& col_id_name_map,
             const std::unordered_map<std::string, ColumnValueRangeType>* colname_to_value_range,
             const VExprContextSPtrs& conjuncts, const TupleDescriptor* tuple_descriptor,
             const RowDescriptor* row_descriptor,
@@ -240,7 +240,7 @@ public:
 
     Status init_reader(
             const std::vector<std::string>& file_col_names,
-            const std::unordered_map<int, std::string>& col_id_name_map,
+            const std::unordered_map<uint64_t, std::string>& col_id_name_map,
             const std::unordered_map<std::string, ColumnValueRangeType>* colname_to_value_range,
             const VExprContextSPtrs& conjuncts, const TupleDescriptor* tuple_descriptor,
             const RowDescriptor* row_descriptor,
